rename package name jass to jaas
