+++ functional-test/src/test/resources/conf/standalone.xml
@@ -397,7 +397,7 @@
         </security-domain>
         <security-domain name="zanata.internal">
           <authentication>
-            <login-module code="org.zanata.security.jass.InternalLoginModule"
+            <login-module code="org.zanata.security.jaas.InternalLoginModule"
               flag="required" />
           </authentication>
         </security-domain>
