+++ functional-test/src/test/resources/conf/standalone_wildfly.xml
@@ -355,7 +355,7 @@
         </security-domain>
         <security-domain name="zanata.internal">
           <authentication>
-            <login-module code="org.zanata.security.jass.InternalLoginModule" flag="required"/>
+            <login-module code="org.zanata.security.jaas.InternalLoginModule" flag="required"/>
           </authentication>
         </security-domain>
         <security-domain name="zanata.openid">
