+++ zanata-overlay/distros/eap-6/standalone/configuration/standalone-zanata.xml
@@ -409,7 +409,7 @@
                 </security-domain>
                 <security-domain name="zanata.internal">
                     <authentication>
-                        <login-module code="org.zanata.security.jass.InternalLoginModule" flag="required"/>
+                        <login-module code="org.zanata.security.jaas.InternalLoginModule" flag="required"/>
                     </authentication>
                 </security-domain>
                 <security-domain name="zanata.openid">
