+++ zanata-overlay/distros/wildfly-9/standalone/configuration/standalone-zanata.xml
@@ -451,7 +451,7 @@
                 </security-domain>
                 <security-domain name="zanata.internal">
                     <authentication>
-                        <login-module code="org.zanata.security.jass.InternalLoginModule" flag="required"/>
+                        <login-module code="org.zanata.security.jaas.InternalLoginModule" flag="required"/>
                     </authentication>
                 </security-domain>
                 <security-domain name="zanata.openid">
