+++ zanata-war/src/test/resources/arquillian/standalone-arquillian-wildfly.xml
@@ -370,7 +370,7 @@
         </security-domain>
         <security-domain name="zanata.internal">
           <authentication>
-            <login-module code="org.zanata.security.jass.InternalLoginModule" flag="required"/>
+            <login-module code="org.zanata.security.jaas.InternalLoginModule" flag="required"/>
           </authentication>
         </security-domain>
         <security-domain name="zanata.jaas">
