+++ zanata-war/src/test/resources/arquillian/standalone-arquillian.xml
@@ -395,7 +395,7 @@
         </security-domain>
         <security-domain name="zanata.internal">
           <authentication>
-            <login-module code="org.zanata.security.jass.InternalLoginModule"
+            <login-module code="org.zanata.security.jaas.InternalLoginModule"
               flag="required" />
           </authentication>
         </security-domain>
